Add support for SphinxScopes to has_many associations only if sphinx_scopes are actually used (for performance reasons)
